Migrate more track DB calls to SQLDelight

+++ app/src/main/java/eu/kanade/domain/track/model/Track.kt
@@ -28,7 +28,7 @@ data class Track(
     }
 }
 
-fun Track.toDbTrack(): DbTrack = DbTrack.create(syncId.toInt()).also {
+fun Track.toDbTrack(): DbTrack = DbTrack.create(syncId).also {
     it.id = id
     it.manga_id = mangaId
     it.media_id = remoteId
