+++ app/src/main/java/eu/kanade/tachiyomi/data/backup/AbstractBackupRestore.kt
@@ -87,7 +87,7 @@ abstract class AbstractBackupRestore<T : AbstractBackupManager>(protected val co
      */
     internal suspend fun updateTracking(manga: Manga, tracks: List<Track>) {
         tracks.forEach { track ->
-            val service = trackManager.getService(track.sync_id)
+            val service = trackManager.getService(track.sync_id.toLong())
             if (service != null && service.isLogged) {
                 try {
                     val updatedTrack = service.refresh(track)
