+++ app/src/main/java/eu/kanade/tachiyomi/data/database/models/Track.kt
@@ -39,8 +39,8 @@ interface Track : Serializable {
     }
 
     companion object {
-        fun create(serviceId: Int): Track = TrackImpl().apply {
-            sync_id = serviceId
+        fun create(serviceId: Long): Track = TrackImpl().apply {
+            sync_id = serviceId.toInt()
         }
     }
 }
