+++ app/src/main/java/eu/kanade/tachiyomi/data/preference/PreferenceKeys.kt
@@ -75,9 +75,9 @@ object PreferenceKeys {
 
     const val autoClearChapterCache = "auto_clear_chapter_cache"
 
-    fun trackUsername(syncId: Int) = "pref_mangasync_username_$syncId"
+    fun trackUsername(syncId: Long) = "pref_mangasync_username_$syncId"
 
-    fun trackPassword(syncId: Int) = "pref_mangasync_password_$syncId"
+    fun trackPassword(syncId: Long) = "pref_mangasync_password_$syncId"
 
-    fun trackToken(syncId: Int) = "track_token_$syncId"
+    fun trackToken(syncId: Long) = "track_token_$syncId"
 }
