+++ app/src/main/java/eu/kanade/tachiyomi/data/track/kitsu/Kitsu.kt
@@ -13,7 +13,7 @@ import kotlinx.serialization.json.Json
 import uy.kohesive.injekt.injectLazy
 import java.text.DecimalFormat
 
-class Kitsu(private val context: Context, id: Int) : TrackService(id) {
+class Kitsu(private val context: Context, id: Long) : TrackService(id) {
 
     companion object {
         const val READING = 1
