+++ app/src/main/java/eu/kanade/tachiyomi/data/track/mangaupdates/dto/Record.kt
@@ -23,7 +23,7 @@ data class Record(
     val latestChapter: Int? = null,
 )
 
-fun Record.toTrackSearch(id: Int): TrackSearch {
+fun Record.toTrackSearch(id: Long): TrackSearch {
     return TrackSearch.create(id).apply {
         media_id = this@toTrackSearch.seriesId ?: 0L
         title = this@toTrackSearch.title?.htmlDecode() ?: ""
