+++ app/src/main/java/eu/kanade/tachiyomi/data/track/model/TrackSearch.kt
@@ -61,8 +61,8 @@ class TrackSearch : Track {
     }
 
     companion object {
-        fun create(serviceId: Int): TrackSearch = TrackSearch().apply {
-            sync_id = serviceId
+        fun create(serviceId: Long): TrackSearch = TrackSearch().apply {
+            sync_id = serviceId.toInt()
         }
     }
 }
